+++ src/App.tsx
@@ -24,7 +24,7 @@ if (
 ) {
   init({
     dsn: process.env.REACT_APP_SENTRY_DSN_URL,
-    environment: process.env.REACT_APP_SENTRY_ENV,
+    environment: process.env.REACT_APP_ENV,
     release: process.env.REACT_APP_SENTRY_RELEASE,
   });
   console.log("Sentry logging is initialized");