+++ src/Contexts/Adaptors/SubstrateHomeAdaptor.tsx
@@ -1,18 +1,8 @@
 import React, { useCallback, useEffect, useState } from "react";
-import { DestinationBridgeContext } from "../DestinationBridgeContext";
 import { HomeBridgeContext } from "../HomeBridgeContext";
 import { useNetworkManager } from "../NetworkManagerContext";
-import {
-  createApi,
-  submitDeposit,
-  getBridgeProposalVotes,
-  VoteStatus,
-} from "./SubstrateApis/ChainBridgeAPI";
-import {
-  IDestinationBridgeProviderProps,
-  IHomeBridgeProviderProps,
-  InjectedAccountType,
-} from "./interfaces";
+import { createApi, submitDeposit } from "./SubstrateApis/ChainBridgeAPI";
+import { IHomeBridgeProviderProps, InjectedAccountType } from "./interfaces";
 
 import { ApiPromise } from "@polkadot/api";
 import {
@@ -23,14 +13,10 @@ import {
 import { TypeRegistry } from "@polkadot/types";
 import { Tokens } from "@chainsafe/web3-context/dist/context/tokensReducer";
 import { BigNumber as BN } from "bignumber.js";
-import { UnsubscribePromise, VoidFn } from "@polkadot/api/types";
+import { VoidFn } from "@polkadot/api/types";
 import { utils } from "ethers";
-import {
-  SubstrateBridgeConfig,
-  getСhainTransferFallbackConfig,
-} from "../../chainbridgeConfig";
+import { SubstrateBridgeConfig } from "../../chainbridgeConfig";
 import { toFixedWithoutRounding } from "../../Utils/Helpers";
-import { Fallback } from "../../Utils/Fallback";
 
 export const SubstrateHomeAdaptorProvider = ({
   children,
@@ -39,7 +25,6 @@ export const SubstrateHomeAdaptorProvider = ({
   const [api, setApi] = useState<ApiPromise | undefined>();
   const [isReady, setIsReady] = useState(false);
   const [accounts, setAccounts] = useState<InjectedAccountType[]>([]);
-  const [address, setAddress] = useState<string | undefined>(undefined);
 
   const {
     homeChainConfig,
@@ -49,8 +34,12 @@ export const SubstrateHomeAdaptorProvider = ({
     homeChains,
     depositAmount,
     setDepositAmount,
+    depositRecipient,
     setDepositRecipient,
     fallback,
+    address,
+    setAddress,
+    analytics,
   } = useNetworkManager();
 
   const [relayerThreshold, setRelayerThreshold] = useState<number | undefined>(
@@ -279,8 +268,15 @@ export const SubstrateHomeAdaptorProvider = ({
                   ]
                     .chainNonces(destinationChainId)
                     .then((response) => {
-                      setDepositNonce(`${response.toJSON()}`);
+                      const depositNonce = `${response.toJSON()}`;
+                      setDepositNonce(depositNonce);
                       setTransactionStatus("In Transit");
+                      analytics.trackTransferInTransitEvent({
+                        address,
+                        recipient,
+                        nonce: parseInt(depositNonce),
+                        amount: depositAmount as number,
+                      });
                     })
                     .catch((error) => {
                       console.error(error);
@@ -342,187 +338,5 @@ export const SubstrateHomeAdaptorProvider = ({
     >
       {children}
     </HomeBridgeContext.Provider>
-  );
-};
-
-export const SubstrateDestinationAdaptorProvider = ({
-  children,
-}: IDestinationBridgeProviderProps) => {
-  const {
-    depositNonce,
-    homeChainConfig,
-    destinationChainConfig,
-    setDepositVotes,
-    depositVotes,
-    tokensDispatch,
-    transactionStatus,
-    setTransactionStatus,
-    depositAmount,
-    depositRecipient,
-    setFallback,
-    fallback,
-  } = useNetworkManager();
-
-  const [api, setApi] = useState<ApiPromise | undefined>();
-
-  const [initiaising, setInitialising] = useState(false);
-  useEffect(() => {
-    // Once the chain ID has been set in the network context, the destination configuration will be automatically
-    // set thus triggering this
-    if (!destinationChainConfig || initiaising || api) return;
-    setInitialising(true);
-    createApi(
-      destinationChainConfig.rpcUrl,
-      destinationChainConfig.rpcFallbackUrls
-    )
-      .then((api) => {
-        setApi(api);
-        setInitialising(false);
-      })
-      .catch(console.error);
-  }, [destinationChainConfig, api, initiaising, transactionStatus]);
-
-  const [listenerActive, setListenerActive] = useState<
-    UnsubscribePromise | undefined
-  >(undefined);
-
-  useEffect(() => {
-    if (api && !listenerActive && depositNonce) {
-      // Wire up event listeners
-      // Subscribe to system events via storage
-      const unsubscribe = api.query.system.events((events) => {
-        console.log("----- Received " + events.length + " event(s): -----");
-        // loop through the Vec<EventRecord>
-        events.forEach((record) => {
-          // extract the phase, event and the event types
-          const { event, phase } = record;
-          const types = event.typeDef;
-          // show what we are busy with
-          console.log(
-            event.section +
-              ":" +
-              event.method +
-              "::" +
-              "phase=" +
-              phase.toString()
-          );
-          console.log(event.meta.documentation.toString());
-          // loop through each of the parameters, displaying the type and data
-          event.data.forEach((data, index) => {
-            console.log(types[index].type + ";" + data.toString());
-          });
-
-          if (
-            event.section ===
-              (destinationChainConfig as SubstrateBridgeConfig)
-                .chainbridgePalletName &&
-            event.method === "VoteFor"
-          ) {
-            setDepositVotes(depositVotes + 1);
-            tokensDispatch({
-              type: "addMessage",
-              payload: {
-                address: "Substrate Relayer",
-                signed: "Confirmed",
-                order: parseFloat(`1.${depositVotes + 1}`),
-              },
-            });
-          }
-
-          if (
-            event.section ===
-              (destinationChainConfig as SubstrateBridgeConfig)
-                .chainbridgePalletName &&
-            event.method === "ProposalApproved"
-          ) {
-            setDepositVotes(depositVotes + 1);
-            setTransactionStatus("Transfer Completed");
-            fallback?.stop();
-          }
-        });
-      });
-      setListenerActive(unsubscribe);
-    } else if (listenerActive && !depositNonce) {
-      const unsubscribeCall = async () => {
-        setListenerActive(undefined);
-      };
-      unsubscribeCall();
-    }
-  }, [
-    api,
-    depositNonce,
-    depositVotes,
-    destinationChainConfig,
-    listenerActive,
-    setDepositVotes,
-    setTransactionStatus,
-    tokensDispatch,
-  ]);
-
-  const initFallbackMechanism = useCallback(async (): Promise<void> => {
-    const srcChainId = homeChainConfig?.chainId as number;
-    const destinationChainId = destinationChainConfig?.chainId as number;
-    const {
-      delayMs,
-      pollingMinIntervalMs,
-      pollingMaxIntervalMs,
-      blockTimeMs,
-    } = getСhainTransferFallbackConfig(srcChainId, destinationChainId);
-    const pollingIntervalMs = Math.min(
-      Math.max(pollingMinIntervalMs, 3 * blockTimeMs),
-      pollingMaxIntervalMs
-    );
-    const fallback = new Fallback(delayMs, pollingIntervalMs, async () => {
-      const res = await getBridgeProposalVotes(
-        api as ApiPromise,
-        srcChainId,
-        destinationChainId,
-        depositRecipient as string,
-        parseInt(depositNonce as string),
-        depositAmount as number
-      );
-      console.log("Proposal votes status", res?.status);
-      switch (res?.status) {
-        case VoteStatus.APPROVED:
-          console.log("Transfer completed in fallback mechanism");
-          setTransactionStatus("Transfer Completed");
-          fallback.stop();
-          return false;
-        case VoteStatus.REJECTED:
-          console.log("Transfer aborted in fallback mechanism");
-          setTransactionStatus("Transfer Aborted");
-          fallback.stop();
-          return false;
-        default:
-          return true;
-      }
-    });
-    fallback.start();
-    setFallback(fallback);
-  }, [
-    api,
-    homeChainConfig,
-    destinationChainConfig,
-    depositRecipient,
-    depositNonce,
-    depositAmount,
-  ]);
-
-  useEffect(() => {
-    console.log({ transactionStatus }); // ToDo: check why get transaction status update several times on the same status
-    if (transactionStatus === "In Transit" && api && !fallback?.initialized())
-      initFallbackMechanism();
-  }, [transactionStatus, api, fallback]);
-
-  return (
-    <DestinationBridgeContext.Provider
-      value={{
-        disconnect: async () => {
-          await api?.disconnect();
-        },
-      }}
-    >
-      {children}
-    </DestinationBridgeContext.Provider>
   );
 };